+++ hypertext/src/validation/attributes.rs
@@ -25,30 +25,28 @@
 //! ```
 #![allow(non_upper_case_globals)]
 #[cfg(feature = "mathml")]
-pub use crate::mathml::MathMlGlobalAttributes;
-use crate::validation::Element;
+pub use super::mathml::MathMlGlobalAttributes;
+use crate::validation::{Attribute, Element};
 #[allow(unused_imports)]
-use crate::validation::{Attribute, AttributeNamespace, AttributeSymbol};
+use crate::validation::{AttributeNamespace, AttributeSymbol};
 
 /// Global HTML attributes.
 ///
 /// This trait must be in scope to use standard HTML attributes such as
 /// [`class`](Self::class) and [`id`](Self::id). This trait is implemented
-/// by every HTML element specified in [`crate::html_elements`].
+/// by every HTML element specified in
+/// [`hypertext_elements`](crate::validation::hypertext_elements).
 ///
 /// # Usage With Custom Elements
 ///
 /// ```
 /// use hypertext::prelude::*;
 ///
-/// mod html_elements {
+/// mod hypertext_elements {
 ///     #![expect(non_camel_case_types)]
 ///
-///     pub use hypertext::html_elements::*;
-///     use hypertext::{
-///         attributes::GlobalAttributes,
-///         validation::{Element, Normal},
-///     };
+///     pub use hypertext::validation::hypertext_elements::*;
+///     use hypertext::validation::{Element, Normal, attributes::GlobalAttributes};
 ///
 ///     pub struct custom_element;
 ///
